+++ src/migrations/Install.php
@@ -21,7 +21,6 @@ class Install extends Migration
             $this->createTable(Plugin::CACHE_TABLE, [
                 'uid'         => $this->string(32)->notNull()->unique(),
                 'url'         => $this->string(255)->notNull(),
-                'body'        => $this->mediumText()->defaultValue(null),
                 'headers'     => $this->text()->defaultValue(null),
                 'tags'        => $this->string(255)->notNull(),
                 'siteId'      => $this->integer(),
@@ -40,7 +39,6 @@ class Install extends Migration
             $this->createTable(Plugin::CACHE_TABLE, [
                 'uid'         => $this->string(40)->notNull()->unique(),
                 'url'         => $this->string(255)->notNull(),
-                'body'        => $this->mediumText()->defaultValue(null),
                 'headers'     => $this->text()->defaultValue(null),
                 'tags'        => 'varchar[]',
                 'siteId'      => $this->integer(),